Feature/validation checks

+++ src/main/java/six/team/backend/controller/NewsController.java
@@ -8,6 +8,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import six.team.backend.PageJsonGen;
+import six.team.backend.dao.UserDAO;
 import six.team.backend.model.News;
 import six.team.backend.store.CommentStore;
 import six.team.backend.store.NewsStore;
@@ -43,23 +44,36 @@ public class NewsController {
     public @ResponseBody ResponseEntity<String> addNews(HttpServletRequest request,HttpServletResponse response, @RequestBody String text) {
         boolean success,exists;
         JSONObject object = new JSONObject();
-        exists = News.checkValidity(request.getHeader("title"));
-        if(exists)
-        {
-            object.put("message","The title already exists");
-            return new ResponseEntity<String>(object.toString(), HttpStatus.OK);
-        }
-        else {
-            success = News.save(request.getHeader("title"), text , request.getHeader("permission"));
-            if (success) {
-                object.put("message", "The news was added successfully");
-                return new ResponseEntity<String>(object.toString(), HttpStatus.valueOf(201));
+
+        UserDAO UD = new UserDAO();
+        String token = request.getHeader("token");
+
+        if(UD.getUserGroupPermissions(UD.getUserGroup(token),"newsadd")){
+            exists = News.checkValidity(request.getHeader("title"));
+            if(exists)
+            {
+                object.put("message","The title already exists");
+                return new ResponseEntity<String>(object.toString(), HttpStatus.OK);
             }
-            else{
+            else {
+                success = News.save(request.getHeader("title"), text , request.getHeader("permission"));
+                if (success) {
+                    object.put("message", "The news was added successfully");
+                    return new ResponseEntity<String>(object.toString(), HttpStatus.valueOf(201));
+                }
+                else{
                     object.put("message", "The news was not added successfully");
                     return new ResponseEntity<String>(object.toString(), HttpStatus.valueOf(401));
                 }
+            }
+
+        }else {
+            JSONObject message = new JSONObject();
+            message.put("user", "You are Unauthorized to view this content");
+            return new ResponseEntity<String>(message.toString(), HttpStatus.UNAUTHORIZED);
         }
+
+
 
     }
 
@@ -81,46 +95,71 @@ public class NewsController {
     @RequestMapping(produces = MediaType.APPLICATION_JSON_VALUE,value ="/{slug}",method = RequestMethod.POST)
     public @ResponseBody ResponseEntity<String> updateNews(HttpServletRequest request,HttpServletResponse response,@PathVariable(value = "slug") String slug , @RequestBody String text){
 
-       boolean success,exists;
-        JSONObject message = new JSONObject();
-        if(News.checkValidity(request.getHeader("title"))){
-            if (News.generateSlug(request.getHeader("title")).equals(slug)) {
-                success=News.update(slug, request.getHeader("title"), text, request.getHeader("permission"));
+        UserDAO UD = new UserDAO();
+        String token = request.getHeader("token");
+
+        if(UD.getUserGroupPermissions(UD.getUserGroup(token), "newsedit")){
+
+            boolean success,exists;
+            JSONObject message = new JSONObject();
+            if(News.checkValidity(request.getHeader("title"))){
+                if (News.generateSlug(request.getHeader("title")).equals(slug)) {
+                    success=News.update(slug, request.getHeader("title"), text, request.getHeader("permission"));
+                    if(success) {
+                        message.put("status", "success");
+                        message.put("slug", slug);
+                        return new ResponseEntity<String>(message.toString(), HttpStatus.OK);
+                    } else {
+                        message.put("status", "success");
+                        return new ResponseEntity<String>(message.toString(), HttpStatus.valueOf(501));
+                    }
+
+                } else
+                    return new ResponseEntity<String>(message.toString(), HttpStatus.valueOf(501));
+            }else
+            {
+                success = News.update(slug, request.getHeader("title"), request.getHeader("text"), request.getHeader("permission"));
                 if(success) {
-                    message.put("status", "success");
+                    message.put("status", success);
                     message.put("slug", slug);
                     return new ResponseEntity<String>(message.toString(), HttpStatus.OK);
                 } else {
-                    message.put("status", "success");
+                    message.put("status", success);
                     return new ResponseEntity<String>(message.toString(), HttpStatus.valueOf(501));
                 }
-
-            } else
-                return new ResponseEntity<String>(message.toString(), HttpStatus.valueOf(501));
-        }else
-        {
-            success = News.update(slug, request.getHeader("title"), request.getHeader("text"), request.getHeader("permission"));
-            if(success) {
-                message.put("status", success);
-                message.put("slug", slug);
-                return new ResponseEntity<String>(message.toString(), HttpStatus.OK);
-            } else {
-                message.put("status", success);
-                return new ResponseEntity<String>(message.toString(), HttpStatus.valueOf(501));
             }
+
+        }else {
+            JSONObject message = new JSONObject();
+            message.put("user", "You are Unauthorized to view this content");
+            return new ResponseEntity<String>(message.toString(), HttpStatus.UNAUTHORIZED);
         }
+
+
     }
 
 
     @RequestMapping(produces = MediaType.APPLICATION_JSON_VALUE,value ="/{slug}/delete",method = RequestMethod.POST)
-    public @ResponseBody ResponseEntity deleteNews(@PathVariable(value = "slug") String slug ){
-        boolean success=News.delete(slug);
-        JSONObject object = new JSONObject();
-        object.put("message","news deleted: " + success);
-        if(success)
-            return new ResponseEntity<String>(object.toString(), HttpStatus.OK);
-        else
-            return new ResponseEntity<String>(object.toString(), HttpStatus.valueOf(501));
+    public @ResponseBody ResponseEntity deleteNews(HttpServletRequest request, @PathVariable(value = "slug") String slug ){
+
+        UserDAO UD = new UserDAO();
+        String token = request.getHeader("token");
+
+        if(UD.getUserGroupPermissions(UD.getUserGroup(token),"newsdelete")){
+
+            boolean success=News.delete(slug);
+            JSONObject object = new JSONObject();
+            object.put("message","news deleted: " + success);
+            if(success)
+                return new ResponseEntity<String>(object.toString(), HttpStatus.OK);
+            else
+                return new ResponseEntity<String>(object.toString(), HttpStatus.valueOf(501));
+
+        }else {
+            JSONObject message = new JSONObject();
+            message.put("user", "You are Unauthorized to view this content");
+            return new ResponseEntity<String>(message.toString(), HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @RequestMapping(produces = MediaType.APPLICATION_JSON_VALUE, value = "{slug}/comments", method = RequestMethod.GET)
@@ -139,10 +178,21 @@ public class NewsController {
     public
     @ResponseBody
     ResponseEntity addComments(HttpServletRequest request, HttpServletResponse response, @PathVariable(value = "slug") String slug) {
-        boolean success = News.addComment(slug, request.getHeader("text"), request.getHeader("author"));
-        JSONObject object = new JSONObject();
-        object.put("message",success);
-        return new ResponseEntity<String>(object.toString(), HttpStatus.OK);
+
+        UserDAO UD = new UserDAO();
+        String token = request.getHeader("token");
+
+        if(UD.userCheck(UD.getUserName(token))){
+            boolean success = News.addComment(slug, request.getHeader("text"), request.getHeader("author"));
+            JSONObject object = new JSONObject();
+            object.put("message",success);
+            return new ResponseEntity<String>(object.toString(), HttpStatus.OK);
+        }else {
+            JSONObject message = new JSONObject();
+            message.put("user", "You are Unauthorized to view this content");
+            return new ResponseEntity<String>(message.toString(), HttpStatus.UNAUTHORIZED);
+        }
+
 
     }
 
@@ -150,21 +200,44 @@ public class NewsController {
     public
     @ResponseBody
     ResponseEntity deleteComment(HttpServletRequest request, HttpServletResponse response, @PathVariable(value = "commentid") String commentid, @PathVariable(value = "slug") String slug) {
-        boolean success = News.deleteComment(Integer.parseInt(commentid));
-        JSONObject object = new JSONObject();
-        object.put("message",success);
-        return new ResponseEntity<String>(object.toString(), HttpStatus.OK);
+
+        UserDAO UD = new UserDAO();
+        String token = request.getHeader("token");
+        String author = request.getHeader("author");
+
+        if(UD.getUserGroupPermissions(UD.getUserGroup(token),"commentsedit") || author.equals(UD.getUserName(token))){
+            boolean success = News.deleteComment(Integer.parseInt(commentid));
+            JSONObject object = new JSONObject();
+            object.put("message",success);
+            return new ResponseEntity<String>(object.toString(), HttpStatus.OK);
+
+        }else {
+            JSONObject message = new JSONObject();
+            message.put("user", "You are Unauthorized to view this content");
+            return new ResponseEntity<String>(message.toString(), HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @RequestMapping(produces = MediaType.APPLICATION_JSON_VALUE, value = "/{slug}/comment/{commentid}", method = RequestMethod.POST)
     public
-    @ResponseBody
-    ResponseEntity editComment(HttpServletRequest request, HttpServletResponse response, @PathVariable(value = "commentid") String commentid, @PathVariable(value = "slug") String slug) {
-        boolean success = News.editComment(Integer.parseInt(commentid), request.getHeader("text"));
-        JSONObject object = new JSONObject();
-        object.put("message",success);
-        return new ResponseEntity<String>(object.toString(), HttpStatus.OK);
-    }
+    @ResponseBody ResponseEntity editComment(HttpServletRequest request, HttpServletResponse response, @PathVariable(value = "commentid") String commentid, @PathVariable(value = "slug") String slug) {
 
+        UserDAO UD = new UserDAO();
+        String token = request.getHeader("token");
+        String author = request.getHeader("author");
+
+        if (UD.getUserGroupPermissions(UD.getUserGroup(token), "commentsedit") || author.equals(UD.getUserName(token))) {
+
+            boolean success = News.editComment(Integer.parseInt(commentid), request.getHeader("text"));
+            JSONObject object = new JSONObject();
+            object.put("message", success);
+            return new ResponseEntity<String>(object.toString(), HttpStatus.OK);
+
+        } else {
+            JSONObject message = new JSONObject();
+            message.put("user", "You are Unauthorized to view this content");
+            return new ResponseEntity<String>(message.toString(), HttpStatus.UNAUTHORIZED);
+        }
+    }
 
 }
